Update to Bryan's I.A and I.B
Made some grammatical changes to I.A and I.B (sentences that ran a little too long, etc.) and tried to clarify a few points.  Made the definition of a qubit a little more precise.  Fixed an error in the definition of 'efficient' (the term 'sub-exponential' has a precise usage that is not considered efficient, for example, the best general factoring algorithm is sub-exponential)